vec_mat: Add vec4f and mat4f helper functions
This allows not using the ugly nested compound literal syntax, and also
enables writing matrices in their logical layout instead of writing
matrix columns as rows in source code.

These helpers are not used in vec_mat.c to ensure the compiler can
inline the inner functions and merge/reorder/vectorize the steps when
optimization is enabled.